+++ src/main/java/com/lucky/around/meal/common/security/util/CookieProvider.java
@@ -45,11 +45,12 @@ public class CookieProvider {
 
   // 로그아웃 시 리프레시 토큰 삭제
   public Cookie deleteRefreshTokenCookie(HttpServletRequest request) {
-
+    // 리퀘스트에서 refreshToken 빼오기
     Optional<Cookie> findCookie =
         Arrays.stream(request.getCookies())
             .filter(cookie -> cookieName.equals(cookie.getName()))
             .findFirst();
+    // 토큰 만료기간 0으로 설정
     if (findCookie.isPresent()) {
       Cookie cookie = findCookie.get();
       cookie.setMaxAge(0);
